requirements.txt
BzrTools==2.6.0
Django==1.9
Pillow==2.3.0
PyYAML==3.12
SecretStorage==2.0.0
argparse==1.2.1
awsebcli==3.11.0
blessed==1.14.2
botocore==1.7.18
bzr==2.7.0.dev1
cement==2.8.2
chardet==2.0.1
colorama==0.3.7
configobj==4.7.2
decorator==3.4.0
django-bootstrap-datepicker==1.2.3
django-bootstrap-form==3.3
django-datetime-widget==0.9.3
django-filter==1.0.4
django-tables2==1.10.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
duplicity==0.6.23
html5lib==0.999
httplib2==0.8
ipython==1.2.1
jmespath==0.9.3
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lockfile==0.8
lpthw.web==1.1
matplotlib==1.3.1
mercurial==2.8.2
mysqlclient==1.3.10
numpy==1.8.2
oauth==1.0.1
paramiko==1.10.1
pathspec==0.5.0
pexpect==3.1
psycopg2==2.7.3
pycrypto==2.6.1
pygobject==3.12.0
pygpgme==0.3
pyparsing==2.0.1
python-apt===0.9.3.5ubuntu2
python-dateutil==2.6.1
pytz==2012rc0
requests==2.9.1
scipy==0.13.3
semantic-version==2.5.0
simplegeneric==0.8.1
simplejson==3.3.1
six==1.5.2
stevedore==0.14.1
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.7.1
virtualenv==1.11.4
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wadllib==1.3.2
wcwidth==0.1.7
websocket-client==0.44.0
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
